+++ docs/versions/change_log.md
@@ -1,5 +1,10 @@
 # Change log
 
+#### `v0.4.0`
+- Modified the aggregate `handle` method to expect a reply of `Self::Error`. 
+Previously `AggregateError` was returned which allowed overloading enum branches other than the `UserError`, this is no longer allowed.
+- Removed deprecated `EventEnvelope` methods and `UserErrorPayload` struct from use in examples.
+
 #### `v0.3.2`
 - Removed deprecated methods and structs from use in examples.